Working test deployment automation
This is the first mostly working version of the exit server installation
script. What it does right now is setup wireguard and babel services and
convert a list of user details created by the firmware script into
various required configuration parameters. It's imdepotent (meaning
you just run the playbook again to make changes as) and not disruptive
to running wireguard connections. In the future we can add a load
balancer and scale up with ease.