Add syslog collection from participating routers
This results in an aggregation structure where logs go into the exits
over unencrypted udp syslog and then get transformed into encrypted
tcp gelf before being passed to the Graylog host. While this can in
theory help with the volume of messages by reducing the number of
active streams it's mostly an effort to secure the logs without
having a super complicated on router setup to do so.