Add semi-persistant keys for exit per hop tunnels
When we refactored for distributed exits we had to seperate the exit
and relay keys in order to solve some nat punching issues, at the time
I setup the exits to simply generate a new key every restart. But this
causes issues with debts tracking over the long term as you have two keys
and two identities in memory.

This solves that problem by storing and templating a seperate different key
for the per hop tunnels on each exit