Add support for creating vlan interfaces
Previously with Centos we manually created the vlan interfaces and just
ensured they where configured properly for peering. Now with this new
configuration script we will create and configure all required interface
definitions from scratch. This requires specific configuration for vlan
interfaces in order to work. Which this patch adds.